@@ -1,25 +1,30 @@
 {
   "name": "@mojaloop/central-services-health",
-  "version": "7.0.0",
+  "version": "7.3.0",
   "description": "Shared code for generic health check servers",
-  "main": "lib/index.js",
-  "types": "lib/index.d.ts",
+  "main": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "files": [
+    "/dist"
+  ],
   "scripts": {
     "test": "npm run test:unit | faucet",
     "pretest": "npm run build",
-    "standard": "standard",
+    "standard": "echo '\\033[1;33m  This project uses eslint instead of standard. Use `npm run lint` instead.'",
+    "lint": "eslint ./src/**/*.ts ./test/**/*.ts",
     "build": "tsc",
     "watch": "tsc -w",
     "test:unit": "ts-node node_modules/tape/bin/tape 'test/unit/**/*.test.ts'",
     "test:xunit": "npm run test:unit | tap-xunit",
     "test:coverage": "nyc --reporter=lcov --require ts-node/register tape test/unit/**/*.test.{ts,js} | tap-spec; nyc report ---reporter=text",
-    "git:stagelib": "git add lib/*ts; git add lib/*.map; git add lib/*.js",
-    "git:standardFix": "standard --fix"
+    "git:lintFix": "npm run lint -- --fix",
+    "audit:resolve": "SHELL=sh resolve-audit",
+    "audit:check": "SHELL=sh check-audit"
   },
   "pre-commit": [
     "build",
-    "test:unit",
-    "git:stagelib"
+    "git:lintFix",
+    "test:unit"
   ],
   "repository": {
     "type": "git",
@@ -40,30 +45,35 @@
   },
   "homepage": "https://github.com/mojaloop/central-services-health#readme",
   "dependencies": {
-    "@hapi/boom": "7.4.2",
     "@hapi/hapi": "18.3.1",
-    "@mojaloop/central-services-error-handling": "5.2.0",
-    "@mojaloop/central-services-shared": "6.4.1"
+    "@mojaloop/central-services-error-handling": "7.3.0",
+    "@mojaloop/central-services-shared": "7.2.0"
   },
   "devDependencies": {
     "@types/hapi__hapi": "18.2.5",
     "@types/node": "11.9.4",
     "@types/sinon": "7.0.13",
     "@types/tape": "4.2.33",
+    "@typescript-eslint/eslint-plugin": "1.11.0",
+    "@typescript-eslint/parser": "1.11.0",
     "debug": "4.1.1",
+    "eslint": "6.0.1",
+    "eslint-config-prettier": "6.0.0",
+    "eslint-plugin-prettier": "3.1.0",
     "faucet": "0.0.1",
-    "nyc": "13.3.0",
+    "npm-audit-resolver": "1.5.0",
+    "nyc": "14.1.1",
     "pre-commit": "1.2.2",
+    "prettier": "1.18.2",
     "sinon": "7.3.2",
-    "standard": "12.0.1",
     "tap-spec": "5.0.0",
     "tap-xunit": "2.3.0",
-    "tape": "4.10.1",
+    "tape": "4.11.0",
     "tapes": "4.1.0",
-    "ts-node": "8.0.2",
+    "ts-node": "8.3.0",
     "tslint": "5.12.1",
-    "typedoc": "0.14.2",
-    "typescript": "3.3.3333"
+    "typedoc": "0.15.0",
+    "typescript": "3.5.2"
   },
   "nyc": {
     "include": [
