@@ -22,10 +22,8 @@
  --------------
  ******/
 
-import Hapi, { Lifecycle } from '@hapi/hapi';
-import Boom from '@hapi/boom';
-
-const ErrorHandling = require('@mojaloop/central-services-error-handling')
+import Hapi, { Lifecycle, ResponseObject } from '@hapi/hapi'
+const ErrorHandler = require('@mojaloop/central-services-error-handling')
 const Logger = require('@mojaloop/central-services-shared').Logger
 const { responseCode, statusEnum } = require('@mojaloop/central-services-shared').HealthCheck.HealthCheckEnums
 
@@ -40,7 +38,7 @@ const { responseCode, statusEnum } = require('@mojaloop/central-services-shared'
  * @returns {async (response, h) => any} handler - the HapiJS compatible handler for the health check
  */
 const defaultHealthHandler = (healthCheck: any): Lifecycle.Method => {
-  return async (_, h) => {
+  return async (_, h): Promise<ResponseObject> => {
     let responseBody
     let code = responseCode.success
     try {
@@ -64,11 +62,8 @@ const defaultHealthHandler = (healthCheck: any): Lifecycle.Method => {
  * @description the failure handler for Hapi. We put this here to make it more testable
  *
  */
-const failAction = async (_request: Hapi.Request, _handler: Hapi.ResponseToolkit, err?: Error) => {
-  if (!err) {
-    throw Boom.boomify(new Error(`Unknown Server Error`))
-  }
-  throw Boom.boomify(err)
+const failAction = async (_request: Hapi.Request, _handler: Hapi.ResponseToolkit, err?: Error | any): Promise<void> => {
+  throw ErrorHandler.Factory.createFSPIOPErrorFromJoiError(err)
 }
 
 /**
@@ -88,7 +83,7 @@ const createHealthCheckServer = async (port: string, healthCheckHandler: Lifecyc
     port,
     routes: {
       validate: {
-        options: ErrorHandling.validateRoutes(),
+        options: ErrorHandler.validateRoutes(),
         failAction: failAction
       }
     }
@@ -105,8 +100,4 @@ const createHealthCheckServer = async (port: string, healthCheckHandler: Lifecyc
   return server
 }
 
-export {
-  createHealthCheckServer,
-  defaultHealthHandler,
-  failAction
-}
+export { createHealthCheckServer, defaultHealthHandler, failAction }
